requirements.txt
python-OBD==0.7.2      # OBD-II communication (includes pint for units)

# ── Optional ──────────────────────────────────────────────────────────────────
# RPi.GPIO>=0.7.1      # Physical GPIO buttons (set GPIO_ENABLED=True in config.py)